+++ Core/ItemControls.cs
@@ -19,7 +19,7 @@ namespace OlibKey.Core
             for (int i = 0; i < VisualTreeHelper.GetChildrenCount(obj); i++)
             {
                 DependencyObject child = VisualTreeHelper.GetChild(obj, i);
-                if (child != null && child is childItem)
+                if (child is childItem)
                 {
                     return (childItem)child;
                 }