+++ Core/SaveAndLoadAccount.cs
@@ -7,26 +7,14 @@ namespace OlibKey.Core
 {
     public class SaveAndLoadAccount
     {
-        public static Database LoadFiles(string directoryLocation, string masterPassword)
-        {
-            var s = File.ReadAllText(directoryLocation);
-
-            XmlSerializer x = new XmlSerializer(typeof(Database));
-
-            using TextReader reader = new StringReader(Encryptor.DecryptString(s, masterPassword));
-            Database database = (Database)x.Deserialize(reader);
-
-            return database;
-        }
+        public static Database LoadFiles(string directoryLocation, string masterPassword) => (Database)new XmlSerializer(typeof(Database)).Deserialize(new StringReader(Encryptor.DecryptString(File.ReadAllText(directoryLocation), masterPassword)));
 
         public static void SaveFiles(Database database, string directoryLocation)
         {
             var writer = new StringWriter();
             new XmlSerializer(typeof(Database)).Serialize(writer, database);
 
-            string s = writer.ToString();
-
-            File.WriteAllText(directoryLocation, Encryptor.EncryptString(s, MainViewModel.MasterPassword));
+            File.WriteAllText(directoryLocation, Encryptor.EncryptString(writer.ToString(), MainViewModel.MasterPassword));
         }
     }
 }