+++ Views/RequireMasterPasswordWindow.xaml.cs
@@ -54,10 +54,7 @@ namespace OlibKey.Views
 
         private void TbMasterPassword_OnKeyDown(object sender, KeyEventArgs e)
         {
-            if (e.Key == Key.Enter)
-            {
-                LoadStorage();
-            }
+            if (e.Key == Key.Enter) LoadStorage();
         }
 
         private void mainWindow_Loaded(object sender, RoutedEventArgs e)